+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/Renderer.java
@@ -520,4 +520,11 @@ public interface Renderer extends PlayerMessage.Target {
    * #STATE_DISABLED}.
    */
   void reset();
+
+  /**
+   * Releases the renderer.
+   *
+   * <p>The renderer must not be used after calling this method.
+   */
+  default void release() {}
 }